+++ src/Mvc/Model/KyteScript.php
@@ -41,14 +41,14 @@ $KyteScript = [
 		],
 
 		'content'		=> [
-			'type'		=> 't',
+			'type'		=> 'lb',
 			'required'	=> false,
 			'date'		=> false,
 		],
 
 		// if js, then obfuscated version
 		'content_js_obfuscated'		=> [
-			'type'		=> 't',
+			'type'		=> 'lb',
 			'required'	=> false,
 			'date'		=> false,
 		],
@@ -66,6 +66,15 @@ $KyteScript = [
 		],
 
 		'obfuscate_js' => [
+			'type'		=> 'i',
+			'required'	=> false,
+			'size'		=> 1,
+			'unsigned'	=> true,
+			'default'	=> 0,
+			'date'		=> false,
+		],
+
+		'include_all' => [
 			'type'		=> 'i',
 			'required'	=> false,
 			'size'		=> 1,