+++ src/Mvc/Model/NavigationItem.php
@@ -12,14 +12,14 @@ $NavigationItem = [
 
 		'element_id'		=> [
 			'type'		=> 's',
-			'required'	=> true,
+			'required'	=> false,
 			'size'		=> 255,
 			'date'		=> false,
 		],
 
 		'element_class'		=> [
 			'type'		=> 's',
-			'required'	=> true,
+			'required'	=> false,
 			'size'		=> 255,
 			'date'		=> false,
 		],